webapp/requirements.txt
streamlit==1.31.0
Pillow==10.2.0
numpy==1.26.4
tensorflow==2.15.0
opencv-python-headless==4.9.0.80
matplotlib==3.8.2
plotly==5.18.0
